Refactoring classes

+++ Sprite.cs
@@ -9,8 +9,8 @@ namespace Top_Down_shooter
 {
     class Sprite
     {
-        public float X { get; set; }
-        public float Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
         public float Angle { get; set; }
         public Bitmap Image { get; set; }
